+++ website/scripts/simulation.js
@@ -1,0 +1,35 @@
+/**
+ * @import {Options} from './options';
+ */
+
+/**
+ * @param {Object} args
+ * @param {Colors} args.colors
+ * @param {Consts} args.consts
+ * @param {LightweightCharts} args.lightweightCharts
+ * @param {Accessor<ChartOption>} args.selected
+ * @param {Signals} args.signals
+ * @param {Utilities} args.utils
+ * @param {Options} args.options
+ * @param {Datasets} args.datasets
+ * @param {WebSockets} args.webSockets
+ * @param {Elements} args.elements
+ * @param {Ids} args.ids
+ * @param {Accessor<boolean>} args.dark
+ */
+export function init({
+  colors,
+  consts,
+  dark,
+  datasets,
+  elements,
+  ids,
+  lightweightCharts,
+  options,
+  selected,
+  signals,
+  utils,
+  webSockets,
+}) {
+  const simulationElement = elements.simulation;
+}